+++ src/main/java/ladder/domain/LadderLine.java
@@ -1,5 +1,7 @@
 package ladder.domain;
 
+import ladder.strategy.MoveStrategy;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
