+++ src/main/java/ladder/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package ladder.domain;
+package ladder.strategy;
 
 @FunctionalInterface
 public interface MoveStrategy {
